requirements.txt
psutil==5.8.0
absl-py==1.0.0
astunparse==1.6.3
av==9.1.1
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
entrypoints==0.3
fastrlock==0.8
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
imageio==2.16.1
imageio-ffmpeg==0.4.5
jupyter==1.0.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
mistune==0.8.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
pywin32==302
PyYAML==6.0
absl-py==1.0.0
astunparse==1.6.3
av==9.1.1
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
entrypoints==0.3
fastrlock==0.8
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
imageio==2.16.1
imageio-ffmpeg==0.4.5
jupyter==1.0.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
mistune==0.8.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
pywin32==302
PyYAML==6.0
requests-oauthlib==1.3.1
rsa==4.8
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tkVideo==0.1
tkvideoplayer==1.3.0
absl-py==1.0.0
astunparse==1.6.3
av==9.1.1
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
entrypoints==0.3
fastrlock==0.8
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
imageio==2.16.1
imageio-ffmpeg==0.4.5
jupyter==1.0.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
mistune==0.8.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
pywin32==302
PyYAML==6.0
requests-oauthlib==1.3.1
rsa==4.8
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tkVideo==0.1
tkvideoplayer==1.3.0
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.14.0
numpy~=1.22.3
scipy~=1.7.3
six~=1.16.0
future~=0.18.2
psutil~=5.8.0